+++ NEWS.md
@@ -35,6 +35,7 @@ New library features
 --------------------
 * `replace(string, pattern...)` now supports an optional `IO` argument to
   write the output to a stream rather than returning a string ([#48625]).
+* `RegexMatch` objects can now be converted to `NamedTuple`s and `Dict`s.
 
 Standard library changes
 ------------------------